Fix GDrive upload file which name might match the one of a folder
Whenever a file is uploaded to GDrive, there is a check for that file
with and without extension, due to Google Docs files having no
extension. This logic now only kicks in whenever the detected
extensionless file is really a Google Doc file, not a folder.

This makes it possible again to upload a file "test.txt" in a folder
that also has a folder called "test"