Proper separation of quota values in APIs
User::getQuota() and User::setQuota() are now reading the exact value as
set in the database instead of interpreting it. This is how it used to
behave.

Util::getUserQuota() is the method that interprets the quota for example
by replacing a default value with the default quota value from the app
config.