[stable10] Disable background scan for home storage/cache
The home cache might have entries with size=-1 that are outside of
"files/" due to encryption or other metadata that generally do not
require the size to be set or accurate.

Instead of adjusting the HomeCache to ignore entries outside of
"files/", let's just disable it completely. The home storage is in the
data folder and must not receive external changes, so it should never
have any entries with size=-1 in "files/".