[stable10] Add class to filter link share email script parameters
The PHP and related JavaScript code for creating a public link share
email don't (at least as far as I can tell) do any form of validation
nor filtering of the four parameters (link, file, to address, and
expiration date) which the share script needs to create the email.

Given that, this class was created to provide an elementary form of
filtering to help avert the HTML injection security vulnerability
identified in https://github.com/owncloud/security-tracker/issues/255.