Change the styling of the active settings navigation menu item
In the ownCloud UI, when a navigation menu is active, it's not always
obvious that it is. There is a subtle difference in how it's rendered,
but I don't feel that it's that clear, nor distinct. This commit updates
it so that it's that much clear to the user as to which element is
currently active.