[stable10] Fix deletion of selected files in trashbin
The file selection process provides the result
which has file name that doesn't match with the
same in the filesystem. This is with reference
to the trashbin. So lets say user tries to selectall,
the value in this.files doesn't give the name
which can be passed to the ajax request to delete
or undelete for later actions. So the purpose
of this change is to check if the selector is
from trashbin, if so adjust the filename as
filename + '.d' + mtime/1000.

Signed-off-by: Sujith H <sharidasan@owncloud.com>