[stable10] Use the userid obtained from checkPassword after login is validated
This change includes:
a) Change the function argument from 'uid' to 'login'. The php
doc is also updated.
b) before event argument would use 'login' to have loginname
c) after event argument would use 'user' instead of 'uid', because
'user' is an instance of IUser. We pass User object.

The user id sent to the after event should use the id obtained
from checkPassword method, which validates the login credentials
passed to loginWithPassword method. This change would help the
the subscriber of the after event to get the appropriate userid.

Signed-off-by: Sujith H <sharidasan@owncloud.com>