Don't remove temp file on fail when creating office preview
Whenever an office file preview is created, in case of failure it would
try to delete the input path. In some cases the input path points to a
temporary file and in some other cases it's the original file. Deleting
the original file results in data loss in case the PDF processor had an
error.

The line that deletes the input file was removed now because there is no
need to remove temporary files because the TempManager will have all
temporary files cleant up in the shutdown handler at the end of the PHP
request.