Test files:scan
This commit, although a bit hacky, adds the test for the `files:scan`
in the Occ. The config didn't seem to have come into effect when set.
But, i noticed that the config was coming into effect when it was
force-scanned. Needs further investigation if `files_external:option`
is the problem or not.