[stable10] Backport of Fix the share permissions evaluation
When user tries to access "Shared with you"
to know the list of files/folders shared with
the user, the share tab for the files or folder
are having an issue. When the group(s) to which
the user belong is excluded from sharing, the
share options appear in the tab. This change
tries to fix the issue.

Signed-off-by: Sujith H <sharidasan@owncloud.com>