Include new options for the license manager
These options are intended to be used in one of the following cases:

If the application can be used freely but it has some parts that require
a license the app should check the license with "startGracePeriod =>
false" and "disableApp => false". This way the app will mostly work with
the open parts without starting the grace period nor having a valid
license.

If the app is intended to be used with a license but has some minor
parts that can be used without it, the app should check the license with
"startGracePeriod => true" and "disableApp => false". The grace period
will start (if possible) allowing the full usage of the app for a
limited time, but once the grace period ends the app won't be disabled,
leaving only the free parts running.

Note that it's the app's responsibility to handle the license state
switch by themselves. The grace period could end anytime, as well as a
new license could be used