Add the getSharesByPath share method to the interface
This method was being used in the files_sharing app even though it was
an implementation detail. The IShareProvider interface requires that
method to be implemented in the providers already, and the current share
manager already has the method implemented. This means that this will
just expose that method through the interface
In addition, the pagination wasn't being used and it wasn't implemented
in the share providers, so it has been removed