fix: require search param to not be empty
The better fix would be to change the `isSearchable` method in
lib/public/Util/UserSearch.php. This would require many changes though
since this method is being used at the Manager level and some users of
the managers expect the old behavior. Ideally the `isSearchable` call
would be moved to the controller / API level.

Co-authored-by: Ilja Neumann <ineumann@owncloud.com>