Fix potential problems with the cache due to removing wrong data
The ICache interface (implemented by the CappedMemoryCache) states that
the clear method must remove the items starting with the prefix passed
as parameter. However, the implementation removes the whole cache.
The current usage seems to work because of this flaw. However, assuming
that "bug" is fixed, the implementation will likely have problems due to
clearing (or not) wrong data.
This commit aims to prevent problems when the CappedMemoryCache adjusts
its code according to the interface.