Expose some license information through the LicenseManager
The exposed information is controlled by the license itself.
By default, the license string won't be exposed, but the license can
change this. Furthermore, additional information can be provided by the
license (number of allowed users, informational links to renew the
license, etc), and this information can be quieried by through the
licenseManager