default pollintervall 30000 msec
The desktop client uses a default poll interval of 30000 msec whenever the capabilities announce something less than 5000.
To get in sync with this behaviour, I suggest to also change the value here to 30000.

A value of 60 also leads to confusion, as users asume the unit is seconds, which it is not.

Suggestion for future: Hardcode the unit in the name of the variable, like. e.g. `pollinterval_msec`