Allow to override LDAP SSL usage and Bind Password via env
This introduces two new env setttings: REVA_LDAP_USESSL and REVA_LDAP_BIND_PASSWORD.
Set REVA_LDAP_USESSL to true to be able to force the usage of LDAPS even
when not running on the standard port (636).
REVA_LDAP_BIND_PASSWORD can be used to override the LDAP bind password.
The purpose of this change is to be able to test against exiting LDAP servers and
the embeded libregraph-idm that is deployed be the default configuration in ocis.